+++ src/wormguides/model/LineageData.java
@@ -6,9 +6,13 @@ public interface LineageData {
 
 	String[] getNames(int time);
 
-	Integer[][] getPositions(int time);
+	Double[][] getPositions(int time);
 
-	Integer[] getDiameters(int time);
+	Double[] getDiameters(int time);
+	
+//	Integer[][] getPositions(int time);
+//
+//	Integer[] getDiameters(int time);
 
 	ArrayList<String> getAllCellNames();
