+++ src/wormguides/model/TableLineageData.java
@@ -50,22 +50,40 @@ public class TableLineageData implements LineageData {
 	}
 
 	@Override
-	public Integer[][] getPositions(int time) {
+	public Double[][] getPositions(int time) {
 		time--;
 		if (time >= getTotalTimePoints() || time < 0)
-			return new Integer[1][3];
+			return new Double[1][3];
 
 		return timeFrames.get(time).getPositions();
 	}
-
+	
 	@Override
-	public Integer[] getDiameters(int time) {
+	public Double[] getDiameters(int time) {
 		time--;
 		if (time >= getTotalTimePoints() || time < 0)
-			return new Integer[1];
+			return new Double[1];
 
 		return timeFrames.get(time).getDiameters();
 	}
+	
+//	@Override
+//	public Integer[][] getPositions(int time) {
+//		time--;
+//		if (time >= getTotalTimePoints() || time < 0)
+//			return new Integer[1][3];
+//
+//		return timeFrames.get(time).getPositions();
+//	}
+
+//	@Override
+//	public Integer[] getDiameters(int time) {
+//		time--;
+//		if (time >= getTotalTimePoints() || time < 0)
+//			return new Integer[1];
+//
+//		return timeFrames.get(time).getDiameters();
+//	}
 
 	public int getTotalTimePoints() {
 		return timeFrames.size();
@@ -75,13 +93,13 @@ public class TableLineageData implements LineageData {
 		Frame frame = new Frame();
 		timeFrames.add(frame);
 	}
-
-	public void addNucleus(int time, String name, int x, int y, int z, int diameter) {
+	
+	public void addNucleus(int time, String name, double x, double y, double z, double diameter) {
 		if (time <= getTotalTimePoints()) {
 			int index = time - 1;
 			Frame frame = timeFrames.get(index);
 			frame.addName(name);
-			Integer[] position = new Integer[] { x, y, z };
+			Double[] position = new Double[] { x, y, z };
 			frame.addPosition(position);
 			frame.addDiameter(diameter);
 
@@ -89,6 +107,20 @@ public class TableLineageData implements LineageData {
 				allCellNames.add(name);
 		}
 	}
+
+//	public void addNucleus(int time, String name, int x, int y, int z, int diameter) {
+//		if (time <= getTotalTimePoints()) {
+//			int index = time - 1;
+//			Frame frame = timeFrames.get(index);
+//			frame.addName(name);
+//			Integer[] position = new Integer[] { x, y, z };
+//			frame.addPosition(position);
+//			frame.addDiameter(diameter);
+//
+//			if (!allCellNames.contains(name))
+//				allCellNames.add(name);
+//		}
+//	}
 
 	public int getFirstOccurrenceOf(String name) {
 		int time = Integer.MIN_VALUE;
@@ -152,17 +184,23 @@ public class TableLineageData implements LineageData {
 
 	public class Frame {
 		private ArrayList<String> names;
-		private ArrayList<Integer[]> positions;
-		private ArrayList<Integer> diameters;
+//		private ArrayList<Integer[]> positions;
+		private ArrayList<Double[]> positions;
+//		private ArrayList<Integer> diameters;
+		private ArrayList<Double> diameters;
 
 		private String[] namesArray;
-		private Integer[][] positionsArray;
-		private Integer[] diametersArray;
+//		private Integer[][] positionsArray;
+//		private Integer[] diametersArray;
+		private Double[][] positionsArray;
+		private Double[] diametersArray;
 
 		private Frame() {
 			names = new ArrayList<String>();
-			positions = new ArrayList<Integer[]>();
-			diameters = new ArrayList<Integer>();
+//			positions = new ArrayList<Integer[]>();
+			positions = new ArrayList<Double[]>();
+//			diameters = new ArrayList<Integer>();
+			diameters = new ArrayList<Double>();
 		}
 
 		/**
@@ -178,8 +216,10 @@ public class TableLineageData implements LineageData {
 		 */
 		private void shiftPositions(int x, int y, int z) {
 			for (int i = 0; i < positions.size(); i++) {
-				Integer[] pos = positions.get(i);
-				positions.set(i, new Integer[] { pos[0] - x, pos[1] - y, pos[2] - z });
+//				Integer[] pos = positions.get(i);
+				Double[] pos = positions.get(i);
+//				positions.set(i, new Integer[] { pos[0] - x, pos[1] - y, pos[2] - z });
+				positions.set(i, new Double[] { pos[0] - x, pos[1] - y, pos[2] - z });
 			}
 		}
 
@@ -187,13 +227,21 @@ public class TableLineageData implements LineageData {
 			names.add(name);
 		}
 
-		private void addPosition(Integer[] position) {
+		private void addPosition(Double[] position) {
 			positions.add(position);
 		}
-
-		private void addDiameter(Integer diameter) {
+		
+		private void addDiameter(Double diameter) {
 			diameters.add(diameter);
 		}
+		
+//		private void addPosition(Integer[] position) {
+//			positions.add(position);
+//		}
+
+//		private void addDiameter(Integer diameter) {
+//			diameters.add(diameter);
+//		}
 
 		private String[] getNames() {
 			if (namesArray == null)
@@ -201,16 +249,27 @@ public class TableLineageData implements LineageData {
 			return namesArray;
 		}
 
-		private Integer[][] getPositions() {
-			positionsArray = positions.toArray(new Integer[positions.size()][3]);
+//		private Integer[][] getPositions() {
+//			positionsArray = positions.toArray(new Integer[positions.size()][3]);
+//			return positionsArray;
+//		}
+		
+		private Double[][] getPositions() {
+			positionsArray = positions.toArray(new Double[positions.size()][3]);
 			return positionsArray;
 		}
 
-		private Integer[] getDiameters() {
+		private Double[] getDiameters() {
 			if (diametersArray == null)
-				diametersArray = diameters.toArray(new Integer[diameters.size()]);
+				diametersArray = diameters.toArray(new Double[diameters.size()]);
 			return diametersArray;
 		}
+		
+//		private Integer[] getDiameters() {
+//			if (diametersArray == null)
+//				diametersArray = diameters.toArray(new Integer[diameters.size()]);
+//			return diametersArray;
+//		}
 
 		public String toString() {
 			String out = "";
